Refactor: 탭메뉴 컴포넌트 공통 컴포넌트로 분리 (재사용성 고려)
