+++ src/shared/constants/ingredient.tsx
@@ -1,0 +1,4 @@
+export const DICTIONARY_TYPE = [
+  { tab: '이름순', value: 'name' as const },
+  { tab: '등급', value: 'level' as const },
+];
